apps/api/pyproject.toml
[project]
name = "gsml-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy>=2.0",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "apscheduler>=3.10",
    "tiktoken>=0.8",
]

[tool.setuptools.packages.find]
include = ["app*"]
